Open in File Manager
Add the files to the repository: In GitKraken, navigate to the repository you created. Right-click on the repository and select "Open in File Manager." Paste the copied files from NetBeans into this directory.

Commit the changes: In GitKraken, click on the "Changes" tab. You should see the newly added files in the list of changes. Enter a commit message and click "Commit."

Push the changes to GitHub: In GitKraken, click on the "Remote" button and select "Push." This will upload the changes to the remote repository on GitHub.